moving multiple_els method to a different file
